refactor: Integrate `Dashboard` into `AuthenticatedLayout`
